src/requirements.txt
certstream==1.12
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
termcolor==2.3.0
tzdata==2023.3
websocket-client==1.6.1
